+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Fixed
+
+-   Fix bug where the `--manual`, `--csv-file`, and other flags were ignored
+    if a config file was in use.
 
 ## [0.4.11] 2019-08-13